data-processing/kafka/consumer/requirements.txt
kafka-python==2.0.2
pyyaml==6.0
